[fuse_ctrl, integration] Fix fuse_ctrl background checks init error (#90)
* [fuse_ctrl, integration] Fix fuse_ctrl background checks init error

This reinstates the background checks in the fuse_ctrl provisioning
smoke tests by fixing the order of registers write to enable the
periodic integrity and consistency checks.

The CHECK_TIMEOUT register needs to be written before
INTEGRITY_CHECK_PERIOD and CONSISTENCY_CHECK_PERIOD or an unintended
timeout error is triggered.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>

* [fuse_ctrl, doc] Explicit integrity/consistency check write order

In the fuse_ctrl an integrity/consistency check is triggered as soon as
either the `INTEGRITY_CHECK_PERIOD` or `CONSISTENCY_CHECK_PERIOD`
registers are written. The following check, depending on the number of
instantiated partitions, may last for hundreds of cycles. Writing a
value into the `CHECK_TIMEOUT` register while the checks are still
ongoing will trigger an unnecessary timeout error that transfers the
fuse_ctlr into an error state.

This commit adjusts the documentation to make it clear that the
`CHECK_TIMEOUT` register should be written before
`INTEGRITY_CHECK_PERIOD` and `CONSISTENCY_CHECK_PERIOD`. Following the
Programmer's Guide verbatim before may lead to implementations that
trigger this unintended error condition.

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>

---------

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>