LCC Tests (#260)
* [lc_ctrl, cov] Fix cov interface and binding

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* [lc_ctrl,test] Add caliptra_ss_lcc_escalation

This test initializes randomly the FC VMEM file and sets a random LC starting
state as well as LC counter. Then, randomly either raise the esc_scrap_state0
or esc_scrap_state1 port. Check, if we are in the ESCALATE LC state. Then
try a state transition to starting state + 1. This should fail and we still
should be in the ESCALATE LC state.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* [lc_ctrl,test] Add smoke_test_lcc_registers

This test reads and writes all registers in the LCC.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* [lc_ctrl,test] Check if state was reached

Before this commit, we just jumped in some test from one LC state
into the other but we did not check whether we actually got into
that state. This commit adds a new helper function in the lc_ctrl
lib that performs the state transition and then reads out the
LC_CTRL state register to check if the transition was successful.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lcc_changes' with updated timestamp and hash after successful run

* [lc_ctrl, test] Change testing priority and register offset

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lcc_changes' with updated timestamp and hash after successful run

---------

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
Co-authored-by: Pascal Nasahl (LOWRISC C I C) <v-pnasahl@microsoft.com>