[lcc,tests] Adapt lcc_registers test (#303)
* [lcc,tests] Adapt lcc_registers test

This commit adapts the following two tests:
- smoke_test_jtag_lcc_registers
- smoke_test_lcc_registers

such that the read only and write only registers are
properly tested. For read only registers, we perform
a read-write-read and check if the last read matches
the first read. For write only registers, we perform
a read-write-read and check if the last read does not
match the write.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* [lcc,test] Fix smoke_test_jtag_lcc_registers

We need to claim the transition if mutex before reading it.
Furthermore, skip the test in the TCL script that sets the
REGWEN of the register. This would prevent us from reading
the flag from the token register.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>

* test update

---------

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
Co-authored-by: Emre Karabulut <emre.karabulut.dde@gmail.com>