Collected lowRISC changes: bundle 1 (#786)
* [fuse_ctrl,dv]: Fix FcLockedPartitionWriteLock_A with scrambling

This assertion is checking that a partition will respond with an error
to any DAI write once the partition has a digest.

Unfortunately, the timing wasn't quite right: if scrambling is enabled
for the partition, the write mechanism in otp_ctrl_dai spends two
extra cycles computing the scrambled address and data before it tries
to write the value and spots that the partition is locked.

Modelling the exact timing carefully is a bit fiddly (because you
can't do ##[my_variable] in SVA) so this cheats slightly by weakening
the assertion. It thus says that if we expect an error it will come
out in between 2 and 4 cycles (even though we could, in theory decide
whether the value should be 2 or 4).

I've also tidied up the assertion a bit to make it easier to read: it
took me a while to decipher, but we may as well make it easier for the
next person.

* [test] Make VPRINTF a more convenient macro

Use the standard trick with "do { ... } while(0)" to make something
that behaves like a single statement (and expects a semicolon).

Now, you can do things like

  if (a) VPRINTF(LOW, "a\n"); else VPRINTF(LOW, "not a\n");

and the if/else will match up correctly.

This commit touches a lot of files because there were quite a few
places where the macro was used without a trailing semicolon.
Fortunately, sed can help:

    for x in $(git grep -l 'VPRINTF[^;]*)$');    \
      do sed -i 's/\(VPRINTF[^;]*)\)$/\1;/' $x;  \
    done

That's not quite enough though! In the other cases, the VPRINTF call
was split across multiple lines or had a comment between the ; and the
newline. There were only 18 of them though, so I've gone through them
manually and checked that everything is terminated properly.

* [lc_ctrl,dv] Tiny improvements in debug message grammar

* [fuse_ctrl,dv] Tidy up smoke_test_fc_mcu_uds_fe_zeroization's prints

The existing macro trickery doesn't actually work (pasting together
with # concatenates the arguments as strings). Annoyingly, there's not
really a nice way to implement variable argument macros in C like this
if you want to support zero arguments (see the new __VA_OPT__
feature).

Rather than depending on a bleeding-edge C compiler, expand out the
LOG_ERROR macro.

* [fuse_ctrl,dv] Correct part. range in caliptra_ss_fuse_ctrl_axi_id.c

This was prompted by issue #769. I've slightly restructured things to
make the code a bit easier to read and also to ensure a failing test
will complete more quickly.

* [fuse_ctrl,dv] Fix build of tests using the partitions array

To find a list of all the tests that mention the partitions[] array,
you can run:

  $ WP=$(git grep -l 'partitions\[' | grep -v libs)

There are 22 of them! But which of those tests mention the partitions
array in the file that gets compiled with the caliptra-rtl Makefile?
Fortunately, not many of them. Most of the tests (such as
caliptra_ss_fuse_ctrl_axi_id) build a test called "smoke_test_mbox"
for the Caliptra core. The following command gets the tests that don't
actually use that trick:

  $ for f in $WP; do grep -q smoke_test_mbox $(dirname $f)/*.yml && continue; echo $(dirname $f); done
  smoke_test_fc_cptra_zeroization
  smoke_test_fc_mcu_uds_fe_zeroization
  smoke_test_fc_ratchet_seed_lock_en

Interestingly, smoke_test_fc_mcu_uds_fe_zeroization doesn't actually
build a binary for the Caliptra core at all. For the other two tests,
ensure that we build the MCU binary first (and, in particular, its
object files). Then add fuse_ctrl_mmap.o to an AUX_OFILES variable
that's passed into the Makefile.

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_1' with updated timestamp and hash after successful run

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_1' with updated timestamp and hash after successful run

---------

Co-authored-by: EMRE KARABULUT <63821295+ekarabu@users.noreply.github.com>
Co-authored-by: Emre Karabulut <ekarabulut@microsoft.com>