lowRISC change bundle 3 (#799)
* [fuse_ctrl,dv] Avoid some repetition in SVA file

* [fuse_ctrl,dv] Fix address used in caliptra_ss_top_sva

The correct address changed a month ago (with commit 795ecaa). Maybe
hard-coding the literal wasn't such a good idea :-)

* [fuse_ctrl,dv] Fix assertion that zeroize marker is always readable

This wasn't quite true: the MCU isn't able to do a DAI write of the
zeroize marker of a secret partition.

Fix things up and add lots of comments to explain what's going
on (because it took me a while to understand).

* [fuse_ctrl,dv] Allow zeroization error responses in SVA

The existing assertion wasn't quite correct: if a zeroization returns
an error, it might not have done any zeroizing!

* [fuse_ctrl,dv] Trivial tweak to make message prettier

* [fuse_ctrl,dv] Fix some types in zeroization smoke tests

This was prompted by issue #767 (which is completely correct). The fix
moves things into the types to make it impossible to make the same
mistake again.

I've also slightly tweaked check_digest when expected_zeroized is
false: if we don't expect to zeroize a fuse, it's still reasonable
for *half* of the fuse to get zeroized.

* [fuse_ctrl,dv] Move check for zeroization result into library code

The basic idea doesn't really change behaviour at all, but avoids all
of the calls to dai_zer needing to follow up with a read / check of
the rdata.

As a bonus, I've also made the tests slightly more "fragile" (so that
they will exit immediately on an error if they are doing a loop)

When getting this working properly, I had to do some work to figure
out whether we should expect partition access to work properly. For
example, smoke_test_fc_mcu_uds_fe_zeroization has a
zeroization_check_unfeasible function which is supposed to try to
zeroize a partition and notice that it doesn't work. Here, I have
strengthened the checks quite a bit, predicting the DAI status we
should expect (so we check we are getting the right error instead of
e.g. the command not reporting a problem).

Similarly, I had to do a reasonable amount of work to make the code
clearer in caliptra_ss_fuse_ctrl_zeroization_reset (after introducing
and fixing an exciting array of bugs).

* [fuse_ctrl,dv] Avoid repetition in zeroization reset test

No functional change, but the code gets slightly shorter and the log
messages are a bit clearer.

* [fuse_ctrl,dv] Further refactoring in zeroization reset test

This now prints out a more informative message that shows exactly what
we were comparing that didn't match.

* [fuse_ctrl,dv] Add a slightly faster way to write an array of fuses

* [fuse_ctrl,dv] Improve prints from zeroization reset test

No big change, but the "step" indices are now consistent (after I'd
messed them up with previous changes).

* [fuse_ctrl,dv] Avoid unnecessary work in zeroization reset test

The prepare_test function writes magic values to the fuses and then
checks that these magic values have indeed arrived. That's not really
all that necessary though (and takes ages!). This commit switches
things so that partitinos after the first written don't bother
checking that the values have been written.

This won't cause a verification gap, but it might make debugging a
failure marginally harder. (I strongly doubt it though)

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_3' with updated timestamp and hash after successful run

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_3' with updated timestamp and hash after successful run

---------

Co-authored-by: Andreas Kurth (LOWRISC C I C) <v-akurth@microsoft.com>
Co-authored-by: Caleb <11879229+calebofearth@users.noreply.github.com>
Co-authored-by: Caleb Whitehead <cwhitehead@microsoft.com>