lowRISC change bundle 6 (#812)
* [VAL] Fix clearing of `SOC_OTP_CTRL_INTERRUPT_STATE`

This register is W1C and has two fields, so we need to write 0x3 to
clear it.

* [fuse_ctrl,dv] Fix invalid state test

This test tries to ask for a transition with a bogus token. It checks
that the dut replies with an error (which it does!), but then checks
that the transition didn't work, by checking that the current state
hasn't changed.

That isn't quite right: lc_ctrl may also change state to
POST_TRANSITION (whether or not the transition succeeded). This is
documented in the repository that defines the IP:

  https://opentitan.org/book/hw/ip/lc_ctrl/doc/theory_of_operation.html#post-transition-handling

Change the test so that it allows either behaviour.

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_6' with updated timestamp and hash after successful run

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_6' with updated timestamp and hash after successful run

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_6' with updated timestamp and hash after successful run

---------

Co-authored-by: Rupert Swarbrick <rswarbrick@lowrisc.org>
Co-authored-by: Rupert Swarbrick (LOWRISC C I C) <v-rswarbrick@microsoft.com>
Co-authored-by: EMRE KARABULUT <63821295+ekarabu@users.noreply.github.com>
Co-authored-by: Emre Karabulut <ekarabulut@microsoft.com>