lowRISC change bundle 7 (#817)
* [lcc,dv] Drop trailing whitespace in lc_ctrl.c

* [lc_ctrl,dv] Tweak read_lc_state to be more careful about repeats

This should have equivalent behaviour, but return 0xff (a known-bad
value), rather than 0x10, when given the state 0xf6543210.

* [lc_ctrl,dv] Change the signature of sw_transition_req

This is inspired by trying to get the code less repetitive. Now, you
pass an array to the function (or null if there is no token).

I've also made the function return whether the transition claimed to
be successful, which allows some prints to be more informative.

Finally, I managed to remove lots of repeated code in
test_all_lc_transitions_no_RMA_no_SCRAP.

* [lc_ctrl,dv] Remove some duplication in lc_ctrl.c

The bodies of sw_transition_req and sw_transition_req_with_expec_error
were essentially the same. Make them actually the same, passing the
"I expect an error" parameter through.

This was driven by porting the "expected failure" version to the new
set of arguments, which is the rest of the commit.

* [lcc, dv] Tidy up how we corrupt a locked partition

This was prompted by issue 807, where the reporter noticed that
caliptra_ss_fuse_ctrl_init_fail had computed a different set of
partitions to look at from the list in fc_lcc_tb_services.

When generating a bug-fix, I realised that we could quite easily make
this a bit more robust, so I have done that in this commit as well.

The only intentional change in behaviour is that
SW_TEST_UNLOCK_PARTITION is now in the list in fc_lcc_tb_services.

* [fuse_ctrl,dv] Fixes to caliptra_ss_fuse_ctrl_init_fail

There were several problems with this test. Firstly, we sometimes
polled the OTP status too early, and could end up seeing a large fault
cause a single partition error before it got mirrored to the other
partitions.

The expected value of the error code register is also modelled wrongly
for an uncorrectable error: the ECC failure that it causes actually
sets the error code to 6 (CHECK_FAIL_ERROR) instead of
3 (MACRO_ECC_UNCORR_ERROR).

This commit also adds some "early exits on failure" to avoid wasting
so much time if something has gone wrong.

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_7' with updated timestamp and hash after successful run

---------

Co-authored-by: Emre Karabulut <ekarabulut@microsoft.com>