[fc,dv] Add missing precondition to FcAxiFilterDaiAccessError_A (#838)
* [fc,dv] Add missing precondition to FcAxiFilterDaiAccessError_A

Because discard_fuse_write stays high for some time after a request,
the previous version of this property wasn't actually true. It was of
the form

    discard ##1 done[->1] |=> error

but that doesn't quite work. If you get a DAI request that is
discarded followed by a good one, then the history looks like:

  1. An AXI write triggers a DAI request that should be discarded.
  2. The filter notices, asserting discard_fuse_write [START A]
  3. Some time later, the operation finishes
  4. It correctly responds with an access error [END A]
  5. On the next cycle, discard_fuse_write is still true. [START B]
  6. discard_fuse_write drops at some point.
  ...
  7. Now there's a good DAI request.
  8. On the next cycle, discard_fuse_write is false (no problem here!)
  9. Some time later, this second operation finishes
 10. The response doesn't have an access error. [END B WITH FAILURE!]

Oh dear...

This commit strengthens the precondition to the form

    dai_req ##1 discard ##0 done[->1] |=> error

With the history above, "START B" doesn't happen on step 5. The
property next starts matching on step 7, but won't match on step
8 (because discard_fuse_write is false), so there is no error.

Thank you, pchunghoon, for the careful description of what was going
on (issue #833).

Note that the ##0 is required after the second term (instead of ##1),
because it's possible to see dai_req_i go high then see
discard_fuse_write_i on the following cycle, with dai_cmd_done_o true
at the same time (followed by an AccessError on the following cycle).

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_tighten_dai_sva' with updated timestamp and hash after successful run

---------

Co-authored-by: EMRE KARABULUT <63821295+ekarabu@users.noreply.github.com>
Co-authored-by: Emre Karabulut <ekarabulut@microsoft.com>