[github] Tidy up / describe "Gen File List" in pre-run-check.yml (#841)
* [github] Tidy up / describe "Gen File List" in pre-run-check.yml

There should be no functional change, but I was staring at this code
after a mysterious CI failure and think it could be rather simpler to
reason about.

There is one *visible* change, that we no longer claim we're printing
the first 5 files. I think printing all the files is intentional, so
let's not claim to do something different!

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_github-hash-check' with updated timestamp and hash after successful run

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_github-hash-check' with updated timestamp and hash after successful run

---------

Co-authored-by: Caleb Whitehead <cwhitehead@microsoft.com>
Co-authored-by: EMRE KARABULUT <63821295+ekarabu@users.noreply.github.com>
Co-authored-by: Emre Karabulut <ekarabulut@microsoft.com>