[lc_ctrl,dv] Test-code cleanups and documentation; fix broken caliptra_ss_lcc_st_trans_scrap_no_ppd_pin (#844)
* [lc_ctrl,dv] Tighten type of read_lc_state

This was already returning a value <= 0xff. Reflect that in the type.

* [lc_ctrl,dv] Fix requested transition in caliptra_ss_lcc_errors

The request being made is also failing, but it's not the request
described in the comment above.

* [lc_ctrl,dv] Get rid of exposed calc_lc_state_mnemonic function

This was only ever used to do stuff that we actually need to package
up anyway.

* [lc_ctrl,dv] Document and improve prints in lcc_initialization

No functional change, but the code should be a bit clearer now, with
some repetition removed and slightly better messages printed.

* [lc_ctrl,dv] Move test_*_RMA_no_SCRAP function out of lc_ctrl.h

This is only used in one test: move it there to tidy things up. Also,
make the test a bit quicker if it has failed.

* [lc_ctrl,dv] Factor out repeated sleep code from lc_ctrl.c

This is actually used all over the place in the DV code, but we may as
well merge the occurrences in one file, at least.

* [lc_ctrl,dv] Document last functions in lc_ctrl.h

Also, tighten the return type of read_lc_counter (since it never
returns anything bigger than 0x1f).

* [lc_ctrl,dv] Re-merge four state transition functions to two

Now the "do we expect an error?" question is handled with an argument
in each pair instead of using different functions.

* [lc_ctrl,dv] Avoid repeating numbers for lifecycle indices

Notice that this adds a rather strange looking line to lc_ctrl.c:

        case MANUF:           state_str = "DEV";             break;

The value 16 is called MANUF in lc_ctrl.h (that defines
lc_state_dec_t) but is called "DEV" in lc_ctrl.c.

Long-term, we should probably pick a name, and I wouldn't be surprised
if the correct answer was "MANUF". I'm being cowardly here though and
this commit intentionally avoids changing the meaning of the code.

* [lc_ctrl,dv] Simplify caliptra_ss_lcc_volatile_unlock

The main change in this commit is to define a new function to perform
"the first part of transition_state", but without following up with a
reset.

But I found this quite confusing to properly understand, so the commit
also adds lots of comments to caliptra_ss_lcc_volatile_unlock.c that
explain what's going on (for the next reader).

* [lc_ctrl,dv] Simplify caliptra_ss_lcc_volatile_unlock_wrong_state

This was motivated by avoiding the use of calc_lc_state_mnemonic
outside of lc_ctrl.c.

The obvious solution is to use transition_state(), which takes the
5-bit state index instead. Unfortunately, the existing test was
explicitly checking that lc_ctrl ended up in PostTransSt before it
injected the reset. I don't think that's an enormous cost for other
users, so let's just plonk that check into the guts of
state_transition instead.

Finally, restructure the test to make the code a bit easier to read.

* [lc_ctrl,dv] Simplify caliptra_ss_lcc_volatile_unlock_wrong_token

This is analogous to the previous commit (that simplified the wrong
state form of the test) and all the same description applies.

* [fc_ctrl,dv] Expand out smoke_test_fc_unlock_transitions

This test doesn't change behaviour, but I expanded out the
code (adding lots of comments) to make it clear what's going on, and
simplify the way the state/token for loop was working: it took me a
few minutes to understand!

This was originally motivated by getting rid of the call to
calc_lc_state_mnemonic and it turns out that transition_state_check
actually does exactly what we need here, so we can just use that.

* [lc_ctrl,dv] Get rid of (unused) calc_lc_state_mnemonic

* [lc_ctrl,dv] Fix caliptra_ss_lcc_st_trans_scrap_no_ppd_pin

This test wasn't really working properly and, more excitingly, didn't
check whether it had worked! Fix both problems, and add detailed
comments that explain what's going on.

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_8' with updated timestamp and hash after successful run

* MICROSOFT AUTOMATED PIPELINE: Stamp 'lowrisc_8' with updated timestamp and hash after successful run

---------

Co-authored-by: EMRE KARABULUT <63821295+ekarabu@users.noreply.github.com>
Co-authored-by: Emre Karabulut <ekarabulut@microsoft.com>